fregata: add a stripped down payload
This is going to be problematic: I can't commit the actual data from
Fregata because the values are production ones and I just don't know
how to encrypt or scramble the values on such large payloads. I looked
into the 'vcr' gem which can filter sensitive data but we'll sort it
out later.