factory: use a shared student API to factor our common values
Since we try and maintain factory-replicas of the respective student
APIs, it makes sense to have a common factory that defines "how we
refer to things" and have two inheriting factories that know "how the
API refers to it". This should help factor our some code and make
maintenance easier.